Implement backup download (#37)
This PR implements the `adync_download_backup` function which fully
completes the `StorjBackupAgent`. However, even though the backup agent
has all the abstract methods defined, the way the download works is not
efficient. There may even be timeout related errors for larger files,
and of course this means that the home assistant instance needs to have
the space to spare. Ideally, we would have a python library that could
create a stream that we receive directly instead of having to write to a
file as a middleman.