Improve setup (#39)
This adds a test for the integration setup as well as makes the setup
align more closely with the Google Drive integration in terms of how we
notify the backup listeners. I've also marked off the
`test-before-setup` item on the quality scale even though no additional
testing is happening. This is because as far as I understand things, any
testing that we would need is already being done in the config flow.