Write stream to temp file before upload (#49)
In the HA OS instance it appears that there is some kind of race
condition happening where the Storj upload tries to run before the Local
back finishes running, meaning that the file that we try to read from
hasn't been written yet. This refactors things so that we write the
stream to a tempfile, meaning that we no longer need to include the
Local backup location when we are using the Storj location.