Migrate to toml config & register pytest mark (#61)
The main point of this PR is to register a custom `is_hassio` mark in
pytest. Since it seems like the `pyproject.toml` has been widely adopted
in the python community as the de-facto configuration file format, I
didn't want to continue modifying `setup.cfg` since it is "softly
deprecated". In order to make the desired configuration change, I first
migrated to the new config file format.

---

## Context for why the mark is used

In #48 I made a change so that when we downloaded the remote backup to
the HomeAssistant instance in order to stream it to the end user, it got
downloaded to a "temporary" file that was in the same directory that the
other system backups would be written to (maybe not strictly necessary,
but was useful for debugging). This location is different depending on
the installation type, and I [made sure there was a
test][comment-on-test] to point it out.

Then in #49 I did some refactoring which included using
`NamedTemporaryFile` along with some mocking to provide something
consistent to test against, and this mocking obscured the different path
that was being used. This is still a problem I'd like to address, but at
least this change resolves the pytest warnings.

[comment-on-test]:
https://github.com/bkjohnson/homeassistant-storj-integration/pull/48/files#r1966689183