Point TMPDIR to new location (#63)
For large backups I get the following error:

```
Traceback (most recent call last):
  File "/usr/src/homeassistant/homeassistant/components/backup/manager.py", line 575, in upload_backup_to_agent
    await self.backup_agents[agent_id].async_upload_backup(
    ...<2 lines>...
    )
  File "/config/custom_components/storj/backup.py", line 86, in async_upload_backup
    await self._client.async_upload_backup(open_stream, backup)
  File "/config/custom_components/storj/api.py", line 117, in async_upload_backup
    fp.write(chunk)
    ~~~~~~~~^^^^^^^
  File "/usr/local/lib/python3.13/tempfile.py", line 499, in func_wrapper
    return func(*args, **kwargs)
OSError: [Errno 28] No space left on device
```

This is despite HomeAssistant running on an SSD that has plenty of
gigabytes to spare. It's possible that HomeAssistant or python is
limiting the amount of space that can be written to the default `/tmp`
directory. By pointing tempfiles to a different location I'm hoping to
get around that limit.