Add storj_uplink package and use it to check satellite (#66)
This will be the first in a series of PRs where the goal is to refactor
the integration so that is uses C bindings for `uplink` rather than
creating a subprocess to call `uplink` directly. This should improve
dependency transparency as well as maintenance.

In this first step, we:

1. Introduce the new `storj_uplink` package
1. Refactor one of the earlier steps in the config flow -
`satellite_is_live()` to use the package instead of calling `uplink`
directly